+++ pysyringe/__init__.py
@@ -1,6 +1,7 @@
 __version__ = "2.0.0"
 
 from pysyringe.container import (
+    AsyncFactoryError,
     Container,
     DuplicateFactoryMethodError,
     Provide,
@@ -11,6 +12,7 @@ from pysyringe.container import (
 from pysyringe.singleton import singleton, thread_local_singleton
 
 __all__ = [
+    "AsyncFactoryError",
     "Container",
     "DuplicateFactoryMethodError",
     "Provide",